--- hub-saml/src/main/java/uk/gov/ida/saml/core/domain/IdaMatchingServiceResponse.java
@@ -1,20 +1,0 @@
-package uk.gov.ida.saml.core.domain;
-
-import org.joda.time.DateTime;
-
-public abstract class IdaMatchingServiceResponse extends IdaMessage implements IdaResponse {
-
-    private String inResponseTo;
-
-    protected IdaMatchingServiceResponse() {
-    }
-
-    public IdaMatchingServiceResponse(String responseId, String inResponseTo, String issuer, DateTime issueInstant) {
-        super(responseId, issuer, issueInstant);
-        this.inResponseTo = inResponseTo;
-    }
-
-    public String getInResponseTo() {
-        return inResponseTo;
-    }
-}
