+++ hub-saml/src/main/java/uk/gov/ida/saml/hub/transformers/outbound/RequestAbstractTypeToStringTransformer.java
@@ -4,7 +4,6 @@ import com.google.inject.Inject;
 import org.opensaml.saml.saml2.core.RequestAbstractType;
 import uk.gov.ida.saml.serializers.XmlObjectToBase64EncodedStringTransformer;
 import uk.gov.ida.saml.core.transformers.outbound.decorators.SamlSignatureSigner;
-import uk.gov.ida.saml.hub.transformers.outbound.decorators.SigningRequestAbstractTypeSignatureCreator;
 
 import java.util.function.Function;
