+++ hub-saml/src/main/java/uk/gov/ida/saml/hub/transformers/outbound/decorators/NoOpSamlAttributeQueryAssertionEncrypter.java
@@ -1,7 +1,7 @@
 package uk.gov.ida.saml.hub.transformers.outbound.decorators;
 
 import org.opensaml.saml.saml2.core.AttributeQuery;
-import uk.gov.ida.saml.core.transformers.outbound.decorators.SamlAttributeQueryAssertionEncrypter;
+import uk.gov.ida.saml.core.domain.SamlAttributeQueryAssertionEncrypter;
 
 public class NoOpSamlAttributeQueryAssertionEncrypter extends SamlAttributeQueryAssertionEncrypter {
     public NoOpSamlAttributeQueryAssertionEncrypter() {
